Moves ExposedPortsParser back to frontend as it is useful for not just plugins-common things.
